refactor(cycles): isolate behavior policy contract
